zero-ed enumerators for simplicity, minor fix on two functions, hidden some debugging prints.

+++ utils/partop/header.h
@@ -10,14 +10,13 @@ int getNumberLANS(topology * top, int index);
 int * getLANS(topology * top, int index);
 double ** getSensorRatesByType(topology * top);
 double * getSensorRatesForOneSensorType(topology * top, int index);
-
 double * getServiceRates(topology * top, int index);
 void * getInfo(topology * top, int index);
 double ** getLANServiceTimesByType(topology * top);
 double * getLANServiceTimesForOneLANType(topology * top, int index);
 int getType(topology * top, int index);
-double ** getActuatorPaths(topology * top);
-double * getActuatorPathsIndex(topology * top, int index);
+int ** getActuatorPaths(topology * top);
+int * getActuatorPathsIndex(topology * top, int index);
 int * getActType(topology * top,int index);
 int getActuatorType(topology * top, int index);
 #endif /* HEADER_H */
